% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlformat.R
\name{adams_all}
\alias{adams_all}
\title{ADAMS all criteria searches URL section for any ADAMS search}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_all(type_chr = NA, start_date = NA, end_date = NA)
}
\arguments{
\item{type_chr}{chr: Type of ADAMS document}
}
\value{
url segment for all criteria searches
}
\description{
ADAMS all criteria searches URL section for any ADAMS search
}
\keyword{Internal}
