% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlformat.R
\name{adams_docket_numbers}
\alias{adams_docket_numbers}
\title{Adams Docket URL section for any ADAMS search}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_docket_numbers(docket_numbers = NA, exclude = FALSE)
}
\arguments{
\item{docket_numbers}{dbl/vector: Docket number (or numbers) to be searched on ADAMS}
}
\value{
url segment for docket searches
}
\description{
Adams Docket URL section for any ADAMS search
}
\keyword{Internal}
