% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{adams_search_tail}
\alias{adams_search_tail}
\title{Last URL section for any ADAMS search}
\format{
An object of class \code{character} of length 1.
}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_search_tail
}
\value{
Ending script for a ADAMS search URL
}
\description{
Last URL section for any ADAMS search
}
\keyword{Internal}
