% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlformat.R
\name{adams_search_term}
\alias{adams_search_term}
\title{ADAMS search term URL section for any ADAMS search}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_search_term(search_term)
}
\arguments{
\item{search_term}{chr: Any search term desired. Default is nothing (i.e., NA)}
}
\value{
script for a ADAMS search URL
}
\description{
ADAMS search term URL section for any ADAMS search
}
\keyword{Internal}
