% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlformat.R
\name{adams_type}
\alias{adams_type}
\title{ADAMS search type URL section for any ADAMS search}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_type(type_chr = NA)
}
\arguments{
\item{type_chr}{chr: Type of ADAMS document}
}
\value{
url segment for type searches
}
\description{
ADAMS search type URL section for any ADAMS search
}
\keyword{Internal}
