% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adv_reactor_dockets_lt_2023}
\alias{adv_reactor_dockets_lt_2023}
\title{Advanced Reactor Docket Pull Prior to 2023}
\format{
tibble of ADAMS search results before Jan. 1 2023
}
\usage{
adv_reactor_dockets_lt_2023
}
\description{
This tibble contains a pre-pulled set of search results from the dockets
found in nrcadams::docket_codex using nrcadams::search_long_docket.
}
\details{
During package develop, github actions experienced frequent connection timeouts
when pulling this entire dataset. By pre-exporting this dataset, the burden
on github actions searches should be reduced.
}
\keyword{datasets}
