% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_list}
\alias{collapse_list}
\title{Collapse List}
\usage{
collapse_list(list_to_collapse)
}
\arguments{
\item{list_to_collapse}{list to be compressed}
}
\value{
vector of listed object
}
\description{
Collapse List
}
\keyword{Internal}
