% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decode_resp}
\alias{decode_resp}
\title{Decodes the REST API response from ADAMS}
\source{
\url{https://adams-search.nrc.gov/assets/APS-API-Guide.pdf}
}
\usage{
decode_resp(resp)
}
\arguments{
\item{resp}{response from REST API}
}
\value{
tibble of search results
}
\description{
Takes the response from the REST API and converts it into a tidy tibble.
}
