% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockets.R
\docType{data}
\name{docket_codex}
\alias{docket_codex}
\title{Codex of Docket Numbers}
\format{
A tibble with columns for project names, company names, docket numbers, and logical flag for NLWR applications.
}
\usage{
docket_codex
}
\value{
tibble of docket numbers
}
\description{
Docket numbers are hard to keep track of so let \code{nrcadams}
keep track of them for you! This exported tibble provides the project name,
company, and docket number for a selection of efforts.
}
\examples{
nrcadams::docket_codex
}
\keyword{datasets}
