% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_from_xml}
\alias{extract_from_xml}
\title{Extract search term from XML results}
\usage{
extract_from_xml(xml_results, search_term)
}
\arguments{
\item{xml_results}{XML query results from ADAMS}

\item{search_term}{Search term}
}
\value{
vector of search term results
}
\description{
Extract search term from XML results
}
\keyword{Internal}
