% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rss.R
\name{format_ML_link}
\alias{format_ML_link}
\title{Make URL from ML number}
\usage{
format_ML_link(results_tbl, ML_vector = `ML Number`)
}
\arguments{
\item{results_tbl}{tibble containing search data}

\item{ML_vector}{The vector name for the document ML number}
}
\description{
Make URL from ML number
}
