% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{hyperlink_file_to_name}
\alias{hyperlink_file_to_name}
\title{Combines ULR with name vector to form a hyperlink}
\usage{
hyperlink_file_to_name(search_tbl, name_vct = Title, url_vct = URL)
}
\arguments{
\item{search_tbl}{Search Tibble}

\item{name_vct}{Name of the column (default is Type)}

\item{url_vct}{Name of the URL column (default is URL)}
}
\value{
search_tbl minus the URL column
}
\description{
Combines ULR with name vector to form a hyperlink
}
