% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlformat.R
\name{lubridate_date}
\alias{lubridate_date}
\title{Formats YMD date string into ADAMS search date using lubridate}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
lubridate_date(date = Sys.Date())
}
\arguments{
\item{date}{chr: The date (ymd)}
}
\value{
date for ADAMS search
}
\description{
Formats YMD date string into ADAMS search date using lubridate
}
\keyword{Internal}
