% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_results_tibble}
\alias{make_results_tibble}
\title{Makes a Formatted Tibble from ADAMS URL Search Results}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
make_results_tibble(adams_url)
}
\arguments{
\item{adams_url}{ADAMS search URL}
}
\value{
vector of search term results
}
\description{
Makes a Formatted Tibble from ADAMS URL Search Results
}
\keyword{Internal}
