% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_dates}
\alias{process_dates}
\title{Process first and last dates.}
\usage{
process_dates(start_date = NA, end_date = NA, days_back = NA)
}
\arguments{
\item{start_date}{chr: The earliest date (ymd) search results should
be returned. Cannot be used with days_back.}

\item{end_date}{chr: The latest date (ymd) search results should be
returned. Cannot be used with days_back.}

\item{days_back}{dbl: Length of time the search extends in days since
the document was published on ADAMS? Default is all time (i.e, NA).
Cannot be used with start_date or end_date.}
}
\value{
list of start and end dates
}
\description{
Process first and last dates.
}
\keyword{Internal}
