% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{remove_noisy_files}
\alias{remove_noisy_files}
\title{Removes files from a targeted search results}
\usage{
remove_noisy_files(search_tbl, target_chr, type_vct = Type, title_vct = Title)
}
\arguments{
\item{search_tbl}{Search Tibble with an ADAMS Type vector and Title vector}

\item{target_chr}{The search type of note}

\item{type_vct}{Name of the Type column (default is Type)}

\item{title_vct}{Name of the Title column (default is Title)}
}
\value{
Search Tibble
}
\description{
Searching for a file type can be difficult due to ADAMS's inconsistent
type structure. This function attempts to "cut through the noise" and
target the search results properly
}
