% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockets.R
\docType{data}
\name{RTR_docket_codex}
\alias{RTR_docket_codex}
\title{Codex of Research and Test Reactor Docket Numbers}
\format{
A tibble with columns for project names, docket numbers.
}
\usage{
RTR_docket_codex
}
\value{
tibble of docket numbers
}
\description{
Docket numbers are hard to keep track of so let \code{nrcadams}
keep track of them for you! This exported tibble provides the project name,
and docket number for a selection of RTRs
}
\examples{
nrcadams::docket_codex
}
\keyword{datasets}
