% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACRS.R
\name{search_ACRS}
\alias{search_ACRS}
\title{Conduct ADAMS Search on ACRS Documents}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
search_ACRS(days_back = 7)
}
\arguments{
\item{days_back}{dbl: Length of time the search extends in days since the document was published on ADAMS? Default is all time (i.e, NA). Cannot be used with start_date or end_date.}
}
\value{
tibble of search results
}
\description{
Conduct ADAMS Search on ACRS Documents
}
