% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACRS.R
\name{search_advanced}
\alias{search_advanced}
\title{Conduct ADAMS Search on ACRS Documents}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
search_advanced(
  search_field = "properties_search_all:!(!(AuthorAffiliation,starts,NRC/ACRS,''),",
  results_tag = "ACRS",
  days_back = 7,
  content_search = FALSE
)
}
\arguments{
\item{search_field}{chr: string with the search query}

\item{results_tag}{chr: string with the results tag}

\item{days_back}{dbl: Length of time the search extends in days since the document was published on ADAMS? Default is all time (i.e, NA). Cannot be used with start_date or end_date.}

\item{content_search}{lgl: Is this a content search or a tag search}
}
\value{
tibble of search results
}
\description{
Conduct ADAMS Search on ACRS Documents
}
