% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_docket}
\alias{search_docket}
\title{Conduct ADAMS Search on Docket Numbers}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
search_docket(
  DocketNumber,
  search_term = NA,
  days_back = NA,
  start_date = NA,
  end_date = NA
)
}
\arguments{
\item{DocketNumber}{dbl/vector: Docket number (or numbers) to be searched on ADAMS}

\item{search_term}{chr: Any search term desired. Default is nothing (i.e., NA)}

\item{days_back}{dbl: Length of time the search extends in days since the document was published on ADAMS? Default is all time (i.e, NA). Cannot be used with start_date or end_date.}

\item{start_date}{chr: The earliest date (ymd) search results should be returned. Cannot be used with days_back.}

\item{end_date}{chr: The latest date (ymd) search results should be returned. Cannot be used with days_back.}
}
\value{
tibble of search results
}
\description{
Conduct ADAMS Search on Docket Numbers
}
\examples{
  nrcadams::docket_codex |>
    dplyr::filter(Company == "ACU") |>
    dplyr::pull(DocketNumber) |>
    nrcadams::search_docket()
}
