% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_long_docket}
\alias{search_long_docket}
\title{Conduct a lengthy search on Docket Numbers}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
search_long_docket(
  DocketNumber,
  search_term = NA,
  number_of_intervals = 5,
  start_date = "2013-1-1"
)
}
\arguments{
\item{DocketNumber}{dbl/vector: Docket number (or numbers) to be searched on ADAMS}

\item{search_term}{chr: Any search term desired. Default is nothing (i.e., NA)}

\item{number_of_intervals}{dbl: The maximum number of searches to be conducted}

\item{start_date}{chr: The earliest date (ymd) search results should be returned.}
}
\value{
tibble of search results
}
\description{
Individual ADAMS searches are limited to 1000 results, which can make
searches that are lengthy difficult. Additionally, long ADAMS searches can
take over 10 seconds to complete which may trigger a connection timeout
on some systems.
}
\details{
To avoid these problems, this wrapper script conducts multiple searches from
today back to a starting point in time. The number of searches is set by the
num_interval. For example, if today is 2020/1/1 and the starting date is
2010/1/1 and number_of_intervals is set to 2, this script will conduct two
searches, one from 2010 to 2015 and the other from 2015 to 2020. These
search results are then combined and output to the user.
}
