% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_ml}
\alias{search_ml}
\title{Conduct ADAMS Search on ML numbers}
\source{
\url{https://adams-search.nrc.gov/assets/APS-API-Guide.pdf}
}
\usage{
search_ml(ML_number)
}
\arguments{
\item{ML_number}{chr/vector: ML number (or numbers) to be searched on ADAMS.
All searches must start with ML but partial ML numbers can be searched for
in addition to full ML numbers.}
}
\value{
tibble of search results
}
\description{
Conduct ADAMS Search on ML numbers
}
\examples{
nrcadams::search_ml(ML_number = c("ML19211C119", "ML20014E642", "ML22179A346"))
}
