% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burning_ears.R
\name{search_undocketed}
\alias{search_undocketed}
\title{Why are my ears burning?}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
search_undocketed(company = "Kairos Power", days_back = 7)
}
\arguments{
\item{company}{chr: Name of the company in nrcadams::docket_codex to be searched for.}

\item{days_back}{dbl: Length of time the search extends in days since the document was published on ADAMS? Default is all time (i.e, NA). Cannot be used with start_date or end_date.}
}
\value{
tibble of search results
}
\description{
This function searches for company mentions that are not tagged to a docket.
}
