% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_values}
\alias{search_values}
\title{Conduct ADAMS Search on Author Affliation and Search Terms}
\source{
\url{https://adams-search.nrc.gov/assets/APS-API-Guide.pdf}
}
\usage{
search_values(
  search_term = NA,
  author_affiliation = NA,
  DocketNumber = NA,
  results_tag = NA,
  days_back = 7,
  skip_int = 0,
  max_returns = 1e+06,
  verbosity = 0,
  adams_key = Sys.getenv("ADAMS")
)
}
\arguments{
\item{search_term}{chr: string with the search query}

\item{author_affiliation}{chr: string with the author affiliation}

\item{DocketNumber}{chr: string with the docket number}

\item{results_tag}{chr: string with the results tag}

\item{days_back}{dbl: Length of time the search extends in days since
the document was published on ADAMS? Default is 7 days.}

\item{skip_int}{dbl: Number of results to skip}

\item{max_returns}{dbl: Maximum number of results to return}

\item{verbosity}{dbl: Level of verbosity for the httr2 request}

\item{adams_key}{chr: ADAMS API key}
}
\value{
tibble of search results
}
\description{
Conduct ADAMS Search on Author Affliation and Search Terms
}
