% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_type}
\alias{simplify_type}
\title{Simplify Type Vector}
\usage{
simplify_type(search_tbl, type_vct = Type)
}
\arguments{
\item{search_tbl}{Search Tibble with an ADAMS Type vector}

\item{type_vct}{Name of the Type column (default is Type)}
}
\value{
Search Tibble
}
\description{
Simplify Type Vector
}
