% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adams_interval}
\alias{adams_interval}
\title{ADAMS search posted date interval URL section for any ADAMS search}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_interval(start_date = NULL, end_date = NULL)
}
\arguments{
\item{start_date}{chr: The earliest date (ymd) search results should be returned.}

\item{end_date}{chr: The latest date (ymd) search results should be returned}
}
\value{
script for a ADAMS search URL
}
\description{
ADAMS search posted date interval URL section for any ADAMS search
}
\keyword{Internal}
