% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adams_search_tail}
\alias{adams_search_tail}
\title{Last URL section for any ADAMS search}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}
}
\usage{
adams_search_tail(content_lgl = TRUE)
}
\arguments{
\item{content_lgl}{Default TRUE.
\itemize{
\item TRUE: Content search supports search keywords.
\item FALSE: Advanced search returns results for greater than 1,000 entries.
}}
}
\value{
Ending script for a ADAMS search URL
}
\description{
Last URL section for any ADAMS search
}
\keyword{Internal}
