% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_docket}
\alias{search_docket}
\title{Conduct ADAMS Search on Docket Numbers}
\source{
\url{https://www.nrc.gov/site-help/developers/wba-api-developer-guide.pdf}

\url{https://adams-search.nrc.gov/assets/APS-API-Guide.pdf}
}
\usage{
search_docket(
  DocketNumber,
  search_term = NA,
  days_back = NA,
  start_date = NA,
  end_date = NA,
  document_type = NA,
  skip_int = 0,
  max_returns = 1e+06,
  verbosity = 0,
  adams_key = Sys.getenv("ADAMS")
)
}
\arguments{
\item{DocketNumber}{dbl/vector: Docket number (or numbers)
to be searched on ADAMS}

\item{search_term}{chr: Any search term desired. Default is nothing
(i.e., NA)}

\item{days_back}{dbl: Length of time the search extends in days since
the document was published on ADAMS? Default is all time (i.e, NA).
Cannot be used with start_date or end_date.}

\item{start_date}{chr: The earliest date (ymd) search results should
be returned. Cannot be used with days_back.}

\item{end_date}{chr: The latest date (ymd) search results should be
returned. Cannot be used with days_back.}

\item{document_type}{chr: Type of ADAMS document, currently unsupported}

\item{max_returns}{dbl: Maximum number of returns to pull when using REST
API.}

\item{verbosity}{dbl: Level of verbosity for REST API requests. Integer 0-3,
with 0 being silent and 3 being most verbose.}

\item{skip_int:}{Number of records to skip in the search.}
}
\value{
tibble of search results
}
\description{
Conduct ADAMS Search on Docket Numbers
}
