% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_public_ADAMS}
\alias{search_public_ADAMS}
\title{Conducts a single ADAMS Search on Docket Numbers}
\source{
\url{https://adams-search.nrc.gov/assets/APS-API-Guide.pdf}
}
\usage{
search_public_ADAMS(
  search_term = NA,
  DocketNumber = NA,
  start_date = NA,
  author_affiliation = NA,
  document_type = NA,
  results_tag = NA,
  end_date = NA,
  skip_int = 0,
  verbosity = 0,
  adams_key = Sys.getenv("ADAMS")
)
}
\arguments{
\item{search_term}{chr: Any search term desired. Default is nothing
(i.e., NA)}

\item{DocketNumber}{dbl/vector: Docket number (or numbers)
to be searched on ADAMS}

\item{start_date}{chr: The earliest date (ymd) search results should
be returned. Cannot be used with days_back.}

\item{end_date}{chr: The latest date (ymd) search results should be
returned. Cannot be used with days_back.}

\item{verbosity}{dbl: Level of verbosity for REST API requests.}

\item{skip_int:}{Number of records to skip in the search.}
}
\value{
tibble of search results
}
\description{
Maximum of 100 results per search, so multiple searches may be needed.
Please use search_docket() for most use cases.
}
